+++ src/pages/dapp/browser/DappBrowser.tsx
@@ -10,10 +10,9 @@ import { addHTTPsPrefixIfNeeded, getCronosEvmAsset } from '../../../utils/utils'
 import PasswordFormModal from '../../../components/PasswordForm/PasswordFormModal';
 import RequestConfirmation from '../components/RequestConfirmation/RequestConfirmation';
 import { UserAsset } from '../../../models/UserAsset';
-import { secretStoreService } from '../../../storage/SecretStoreService';
+import { secretStoreService } from '../../../service/storage/SecretStoreService';
 import { Dapp, DappBrowserIPC } from '../types';
 import { ProviderPreloadScriptPath } from './config';
-import packageJson from '../../../../package.json';
 import { walletService } from '../../../service/WalletService';
 import { useRefCallback } from './useRefCallback';
 import { useWebInfoProvider } from './useWebInfoProvider';
@@ -24,7 +23,6 @@ import {
 } from './useWebviewStatusInfo';
 import { LEDGER_WALLET_TYPE } from '../../../service/LedgerService';
 import ErrorModalPopup from '../../../components/ErrorModalPopup/ErrorModalPopup';
-
 
 // use **only** one of the following
 // priority: dapp > dappURL
@@ -242,7 +240,7 @@ const DappBrowser = forwardRef<DappBrowserRef, DappBrowserProps>((props: DappBro
 
   useIPCProvider({
     webview: webviewRef.current,
-    onRequestAddress: (onSuccess,) => {
+    onRequestAddress: onSuccess => {
       // TODO: !! cronosAsset may not be ready
       onRequestAddress.current(onSuccess);
     },
@@ -273,7 +271,7 @@ const DappBrowser = forwardRef<DappBrowserRef, DappBrowserProps>((props: DappBro
       // no-op for now
     },
     onFinishTransaction: (error?: string) => {
-      onFinishTransaction.current(error ?? "");
+      onFinishTransaction.current(error ?? '');
     },
   });
 
@@ -329,7 +327,7 @@ const DappBrowser = forwardRef<DappBrowserRef, DappBrowserProps>((props: DappBro
           title={t('general.passwordFormModal.title')}
           visible
           successButtonText={t('general.continue')}
-          confirmPassword={false}
+          skipRepeatValidation
         />
       )}
       {txEvent && requestConfirmationVisible && (
@@ -347,10 +345,10 @@ const DappBrowser = forwardRef<DappBrowserRef, DappBrowserProps>((props: DappBro
           onConfirm={({ gasLimit, gasPrice }) => {
             setRequestConfirmationVisible(false);
             confirmPasswordCallback?.successCallback({
-              signature: "",
+              signature: '',
               gasLimit,
               gasPrice,
-              decryptedPhrase
+              decryptedPhrase,
             });
           }}
           onCancel={() => {
@@ -363,8 +361,10 @@ const DappBrowser = forwardRef<DappBrowserRef, DappBrowserProps>((props: DappBro
       <webview
         preload={ProviderPreloadScriptPath}
         ref={webviewRef}
-        // useragent is required for some dapps to auto connect, eg. cronoschimps
-        useragent={`Mozilla/5.0 (Linux; Android 8.0.0; Desktop Wallet Build/${packageJson.version}) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/62.0.3202.84 Mobile Safari/537.36`}
+        useragent={window.navigator.userAgent.replace(
+          'chain-desktop-wallet',
+          'Desktop Wallet Build',
+        )}
         style={{
           width: '100%',
           height: 'calc(100vh - 48px)',