+++ src/pages/dapp/components/RequestConfirmation/RequestConfirmation.tsx
@@ -7,6 +7,7 @@ import { useTranslation } from 'react-i18next';
 import './RequestConfirmation.less';
 
 import { useRecoilState } from 'recoil';
+import { ethers } from 'ethers';
 import {
   AssetMarketPrice,
   getAssetAmountInFiat,
@@ -18,7 +19,7 @@ import { Session } from '../../../../models/Session';
 import { SupportedChainName, SUPPORTED_CURRENCY } from '../../../../config/StaticConfig';
 import { Dapp, DappBrowserIPC } from '../../types';
 
-import { middleEllipsis, hexToUtf8, getAssetBySymbolAndChain } from '../../../../utils/utils';
+import { middleEllipsis, hexToUtf8, getAssetBySymbolAndChain, isUnlimited } from '../../../../utils/utils';
 import { walletService } from '../../../../service/WalletService';
 import { walletAllAssetsState } from '../../../../recoil/atom';
 import { useLedgerStatus } from '../../../../hooks/useLedgerStatus';
@@ -36,7 +37,7 @@ interface RequestConfirmationProps {
   wallet: Wallet;
   visible: boolean;
   dapp?: Dapp;
-  onConfirm: (info: { gasPrice: BigNumber, gasLimit: BigNumber }) => void;
+  onConfirm: (info: { gasPrice: BigNumber, gasLimit: BigNumber, event: DappBrowserIPC.Event }) => void;
   onCancel: () => void;
 }
 
@@ -100,9 +101,8 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
           }} />
           <div className="row">
             <div className="title">{t('dapp.requestConfirmation.total.title')}</div>
-            <div>{`${scaledAmount(total.toString(), cronosAsset?.decimals ?? 1)} ${
-              cronosAsset?.symbol
-            }`}</div>
+            <div>{`${scaledAmount(total.toString(), cronosAsset?.decimals ?? 1)} ${cronosAsset?.symbol
+              }`}</div>
           </div>
         </>
       );
@@ -165,10 +165,10 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
           }} />
           <div className="row">
             <div className="title">{t('dapp.requestConfirmation.total.title')}</div>
-            <div>{`${scaledAmount(total.toString(), cronosAsset?.decimals ?? 1)} ${
-              cronosAsset?.symbol
-            }`}</div>
+            <div>{`${scaledAmount(total.toString(), cronosAsset?.decimals ?? 1)} ${cronosAsset?.symbol
+              }`}</div>
           </div>
+
           <div className="row">
             <div className="title">{t('dapp.requestConfirmation.contractAddress.title')}</div>
             <a onClick={() => setIsContractAddressReview(!isContractAddressReview)}>
@@ -179,6 +179,12 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
             className="contract-address"
             hidden={!isContractAddressReview}
           >{`${contractAddress}`}</div>
+          {
+            event.name === 'tokenApproval' && <div className="row">
+              <div className="title">{t('settings.revoke.amount')}</div>
+              <div>{isUnlimited(ethers.BigNumber.from(event.object.amount)) ? `${t('settings.revoke.unlimited')} ${event.object.tokenData.symbol}` : `${scaledAmount(event.object.amount, Number(event.object.tokenData.decimals))} ${event.object.tokenData.symbol}`}</div>
+            </div>
+          }
         </>
       );
     }
@@ -203,11 +209,11 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
       );
       const totalValue =
         assetMarketData &&
-        assetMarketData.price &&
-        currentAsset?.mainnetSymbol === assetMarketData.assetSymbol
+          assetMarketData.price &&
+          currentAsset?.mainnetSymbol === assetMarketData.assetSymbol
           ? `${SUPPORTED_CURRENCY.get(assetMarketData.currency)?.symbol}${numeral(
-              getAssetAmountInFiat(totalScaledAmount, assetMarketData),
-            ).format('0,0.00')} ${assetMarketData?.currency}`
+            getAssetAmountInFiat(totalScaledAmount, assetMarketData),
+          ).format('0,0.00')} ${assetMarketData?.currency}`
           : `${SUPPORTED_CURRENCY.get(currentSession.currency)?.symbol}--`;
 
       setMessage(`${totalScaledAmount} ${currentAsset?.symbol}`);
@@ -227,7 +233,7 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
         }),
       );
 
-      setSubMessage(`${dapp?.url}`);
+      setSubMessage(`${dapp?.url ?? ''}`);
       const asset = getAssetBySymbolAndChain(
         allAssets,
         event.object.tokenData.symbol,
@@ -278,7 +284,8 @@ const RequestConfirmation = (props: RequestConfirmationProps) => {
               onClick={() => {
                 onConfirm({
                   gasLimit,
-                  gasPrice
+                  gasPrice,
+                  event
                 })
               }}
               disabled={